+++ src/backend/InvenTree/stock/api.py
@@ -1290,10 +1290,14 @@ class StockList(
     search_fields = [
         'serial',
         'batch',
+        'location__name',
         'part__name',
         'part__IPN',
         'part__description',
-        'location__name',
+        'supplier_part__SKU',
+        'supplier_part__supplier__name',
+        'supplier_part__manufacturer_part__MPN',
+        'supplier_part__manufacturer_part__manufacturer__name',
         'tags__name',
         'tags__slug',
     ]
@@ -1553,9 +1557,7 @@ class StockTrackingList(
                 deltas = item['deltas'] or {}
 
                 if key in deltas:
-                    item['deltas'][f'{key}_detail'] = related_data.get(
-                        deltas[key], None
-                    )
+                    item['deltas'][f'{key}_detail'] = related_data.get(deltas[key])
 
         if page is not None:
             return self.get_paginated_response(data)